doc/requirements.txt
mkdocs
mkdocs-exclude
mkdocs-material
mkdocs-material-extensions
mkdocs-markdownextradata-plugin
mkdocs-awesome-pages-plugin
pymdown-extensions
mdx_truly_sane_lists
